requirements.txt
bs4==0.0.1
scrapy-crawlera==1.7.0
psycopg2==2.8.5